% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{as.DESeqDataSet}
\alias{as.DESeqDataSet}
\alias{as.DESeqDataSet.ExprDataSet}
\title{Function to coerce a \code{ExprDataSet} object to a \code{DESeqDataSet} instance.}
\usage{
as.DESeqDataSet(object)

as.DESeqDataSet(object)
}
\arguments{
\item{object}{A \code{ExprDataSet} instance.}

\item{...}{additional parameter}

\item{object}{A \code{ExprDataSet} instance.}

\item{...}{additional parameter}
}
\value{
A \code{DESeqDataSet} instance.

A \code{DESeqDataSet} instance.
}
\description{
Coerce to a \code{DESeqDataSet} instance

Coerce to a \code{DESeqDataSet} instance
}
\author{
Wubing Zhang

Wubing Zhang
}
