% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{as.gallery}
\alias{as.gallery}
\title{Convert a list of plots to a gallery}
\usage{
as.gallery(x, ...)
}
\arguments{
\item{x}{A ggplot, gallery, plots collection, or nested plain list of them.}

\item{...}{Passed on to methods.}
}
\value{
A gallery.
}
\description{
Plain lists become galleries recursively. Data frames are not treated as
lists of plots. A plots collection becomes a gallery whose folders are the
slashes in its plot names.
}
\examples{
library(ggplot2)
invisible(as.gallery(list(scatter = ggplot(mtcars, aes(wt, mpg)) + geom_point())))

}
