% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxView.R
\docType{methods}
\name{BoxView}
\alias{BoxView}
\title{Boxplot}
\usage{
BoxView(gg, ctrl, treat, label.c = "Control", label.t = "Treatment",
  ylab = "Expression", xlab = NULL, colab = NULL, flip = FALSE,
  size = 1, main = NULL, filename = NULL, width = 5, height = 4,
  ...)
}
\arguments{
\item{gg}{Data frame}

\item{ctrl}{Index of control samples}

\item{treat}{Index of treat samples}

\item{label.c}{Label of control box}

\item{label.t}{Label of treatment box}

\item{ylab}{Label of y-axis}

\item{xlab}{Label of x-axis}

\item{colab}{Label of color}

\item{flip}{Whether flip coordinate}

\item{size}{Size of box and jitter}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Boxplot for each column
}
\author{
Wubing Zhang
}
