% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOrgGeneList.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{Basic function to convert human to mouse gene names}
\usage{
convertHumanGeneList(x)
}
\arguments{
\item{x}{A vector of gene symbols.}
}
\description{
Basic function to convert human to mouse gene names
}
