% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOrgGeneList.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Basic function to convert mouse to human gene names}
\usage{
convertMouseGeneList(x, idType = "Symbol",
  ref = "~/Jobs/Archive/GeneLists/GeneAnnotation/HOM_MouseHumanSequence.txt")
}
\arguments{
\item{x}{A vector of genes.}

\item{idType}{Entrez Or Symbol}

\item{ref}{Reference table}
}
\description{
Basic function to convert mouse to human gene names
}
