% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrView.R
\docType{methods}
\name{corrView}
\alias{corrView}
\title{Compute correlation of between expression of two genesets}
\usage{
corrView(mat, geneset1, geneset2, xlab = geneset1[1], ylab = geneset2[1])
}
\arguments{
\item{mat}{Matrix like object.}

\item{geneset1}{Gene set 1 which should be row names of mat.}

\item{geneset2}{Gene set 1 which should be row names of mat.}

\item{xlab}{Character specifying the label of x axis.}

\item{ylab}{Character specifying the label of y axis.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned
and further customized.
}
\description{
Compute correlation of between expression of two genesets
}
\author{
Wubing Zhang
}
