% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalyze.R
\docType{methods}
\name{DEAnalyze}
\alias{DEAnalyze}
\title{Differential expression analysis}
\usage{
DEAnalyze(obj, option = "limma", SampleAnn = NULL,
  type = "Microarray")
}
\arguments{
\item{obj}{Matrix like object or an ExprDataSet instance.}

\item{option}{"limma", "DESeq2", or others.}

\item{SampleAnn}{Matrix like object (only when obj is a matrix),
the first column should match colnames in obj, and the second column should be condition}

\item{type}{"Microarray" or "RNASeq", only needed when obj is not ExprDataSet instance.}
}
\value{
An ExprDataSet instance.
}
\description{
Differential expression analysis
}
\seealso{
\code{\link{ExprDataSet-class}}
}
\author{
Wubing Zhang
}
