% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotHeatView.R
\docType{methods}
\name{dotHeatView}
\alias{dotHeatView}
\title{Dot heatmap}
\usage{
dotHeatView(mat, limit = c(-2, 2))
}
\arguments{
\item{mat}{Matrix like object, each row is gene and each column is sample.}

\item{limit}{Two-length vector.}
}
\value{
An object created by \code{ggplot}, which can be assigned
and further customized.
}
\description{
Dot heatmap
}
\author{
Wubing Zhang
}
