% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichHGT.R
\docType{methods}
\name{enrichHGT}
\alias{enrichHGT}
\alias{Hypergeometric}
\title{Do enrichment analysis using Hypergeometric test}
\usage{
enrichHGT(geneList, universe = NULL, type = "KEGG", organism = "hsa",
  pvalueCutoff = 0.25, pAdjustMethod = "BH", minGSSize = 2,
  maxGSSize = 500, gmtpath = NA)
}
\arguments{
\item{geneList}{A gene list.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{type}{Geneset category for testing, KEGG(default) or GMT (gmtpath is required).}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{minGSSize}{Minimal size of each geneSet for testing.}

\item{maxGSSize}{Maximal size of each geneSet for analyzing.}

\item{gmtpath}{The file path of GMT genesets.}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using Hypergeometric test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:100]
enrichRes <- enrichHGT(genes)
head(enrichRes@result)

}
\seealso{
\code{\link{enrich.GOstats}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Feizhen Wu
}
