% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ExprDataSet-class}
\alias{ExprDataSet-class}
\alias{ExprDataSet}
\alias{show,ExprDataSet-method}
\alias{summary,ExprDataSet-method}
\title{Class "ExprDataSet"
This class represents the datasets of expression profiles.}
\description{
Class "ExprDataSet"
This class represents the datasets of expression profiles.
}
\section{Slots}{

\describe{
\item{\code{profile}}{Expression profile.}

\item{\code{normalizedDS}}{Normalized dataset.}

\item{\code{SampleAnn}}{For matrix input: a DataFrame with at least two columns.
`sampleName` and `condition`.}

\item{\code{geneTable}}{A DataFrame with gene annotations.}

\item{\code{type}}{"Microarray" or "RNASeq".}

\item{\code{GPL}}{GPL ID.}

\item{\code{Process}}{Description of normalization process.}

\item{\code{DEGRes}}{Summary result of differential expressed genes.}
}}

\author{
WubingZhang
}
\keyword{classes}
