% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractReads.R
\docType{methods}
\name{extractReads}
\alias{extractReads}
\title{Extract reads from fastq file}
\usage{
extractReads(fastq_file)
}
\arguments{
\item{fastq_file}{Path to fastq file.}
}
\value{
A character vector, in which each item is a read sequence.
}
\description{
Extract reads from fastq file
}
\author{
Wubing Zhang
}
