% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterN.R
\name{filterN}
\alias{filterN}
\title{Filter out rows with NA or low value}
\usage{
filterN(m, design = NULL, minS = 3, out = c(NA, 0),
  imputeNA = TRUE, ...)
}
\arguments{
\item{m}{A matrix-like object.}

\item{design}{A vector, indicating the group of columns and specifying to
filter out low-quality rows within each condition.}

\item{minS}{A numeric, specifying the minimum proportion/number of values should
be quantified for each row.}

\item{out}{A vector, specifying low-quality values (NA and 0).}
}
\value{
A matrix with the same columns as input matrix.
}
\description{
Filter out rows with NA or low value
}
\author{
Wubing Zhang
}
