% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformCount.R
\name{filterNvoom}
\alias{filterNvoom}
\title{Filtering a count matrix to remove low read genes}
\usage{
filterNvoom(m, minSamples = ncol(m)/10, minCpm = 0.25)
}
\arguments{
\item{m}{raw count matrix (genes are rows, columns are samples)}

\item{minSamples}{numeric; minimum number of samples in which CPM threshold has to be reached; defaults to a tenth of available samples}

\item{minCpm}{numeric; the minimum counts per million (CPM) that has to be reached}
}
\value{
filtered count matrix
}
\description{
This function takes a count matrix and removes genes that don't reach a minimum number of reads in a certain number of samples.
}
\author{
Dorothee Nickles
}
