% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTILCorrView.R
\docType{methods}
\name{GeneTILCorrView}
\alias{GeneTILCorrView}
\title{Dot heatmap}
\usage{
GeneTILCorrView(gene, cibersort = "TCGA_CIBERSORT.rds",
  exprdir = "ProcessedDat/")
}
\arguments{
\item{gene}{A vector of gene names.}

\item{cibersort}{The path to cibersort Rdataset.}

\item{exprdir}{The directory of processed TCGA expression profiles.}
}
\value{
An object created by \code{ggplot}, which can be assigned
and further customized.
}
\description{
Dot heatmap
}
\author{
Wubing Zhang
}
