% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getDEG}
\alias{getDEG}
\title{Get summary of differential expression results}
\usage{
getDEG(x)
}
\arguments{
\item{x}{ExprDataSet object}
}
\value{
A data frame which include differential expression analysis results
}
\description{
Get summary of differential expression results
}
