% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggview.R
\name{ggview}
\alias{ggview}
\title{Ggplot picture previewer}
\usage{
ggview(
  plot = ggplot2::last_plot(),
  device = c("png", "jpeg", "bmp", "svg"),
  scale = 1,
  width,
  height,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  bg = NULL
)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{device}{Device to use. Can be one of "png", "jpeg", "bmp" or "svg".}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
}
\description{
Preview what a ggplot would look like is you save it to a file.
Set picture parameters as you would set them in \link[ggplot2:ggsave]{ggplot2::ggsave} and see
the result in RStudio viewer.
}
\examples{
\dontrun{
library(ggplot2)
p <-
  ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  ggtitle("My awesome plot")

ggview(width = 3, height = 3)
ggview(width = 5, height = 3)
}

}
