% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatPlot.R
\name{HeatmapJView}
\alias{HeatmapJView}
\title{Typical heatmap from Jingxin}
\usage{
HeatmapJView(dat, x, y, rho, pvalue, rho.limit = c(-1, 0, 1),
  bar_anno = c(-1, 0, 1), box_size = 9, barwidth = 0.8,
  barheight = 5)
}
\arguments{
\item{dat}{Data frame.}

\item{x}{The variable of x.}

\item{y}{The variable of y.}

\item{rho}{The variable of color.}

\item{pvalue}{The pvalue column.}

\item{rho.limit}{A numberic vector with three values.}

\item{bar_anno}{Limit on lengend.}

\item{box_size}{A numberic value to specifying the box size.}

\item{barwidth}{A numberic value to specifying the width of colorbar.}

\item{barheight}{A numberic value to specifying the height of colorbar.}
}
\value{
ggplot object
}
\description{
Typical heatmap from Jingxin
}
\author{
Jingxin Fu
}
