% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KcHeatView.R
\docType{methods}
\name{KcHeatView}
\alias{KcHeatView}
\title{K-means clustering and plot Heatmap}
\usage{
KcHeatView(dd, limit = 2, cluster_row = TRUE, cluster_col = TRUE, k = 5,
  filename = NA, height = 0.12 * nrow(dd) + 3, width = 0.12 * ncol(dd) +
  7, ...)
}
\arguments{
\item{dd}{Matrix like object.}

\item{cluster_row}{Boolean, whether cluster rows.}

\item{cluster_col}{Boolean, whether cluster columns.}

\item{k}{Integer, how many clusters to be clustered.}

\item{filename}{Character, specifying output pdf name.}

\item{height}{Integer, sepcifying height of heatmap.}

\item{width}{Integer, sepcifying width of heatmap.}

\item{...}{Other available parameters for pheatmap}
}
\value{
Invisibly a pheatmap object.
}
\description{
K-means clustering and plot Heatmap
}
\author{
Wubing Zhang
}
