% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucleoTranslate.R
\docType{methods}
\name{NucleoTranslate}
\alias{NucleoTranslate}
\title{Nucleotide Sequence Translation
Translate nucleic acid sequence to corresponding peptide sequences.}
\usage{
NucleoTranslate(sequence)
}
\arguments{
\item{sequence}{A character, specifying the nucleic acid sequence.}
}
\value{
A character of peptide sequence.
}
\description{
Nucleotide Sequence Translation
Translate nucleic acid sequence to corresponding peptide sequences.
}
\examples{
sequence = "AATGAGGTTGGTGCGGAGTTTCATCCTACTATGTTG"
NucleoTranslate(sequence)
}
\author{
Wubing Zhang
}
