% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieGridView.R
\name{PieGridView}
\alias{PieGridView}
\title{A matrix of pie plots}
\usage{
PieGridView(
  dat,
  y = "Freq",
  fill = "Note",
  rows = "Gene",
  cols = "Cancer",
  percentage = TRUE,
  size = 2,
  row.label = NULL,
  col.label = NULL
)
}
\arguments{
\item{dat}{A data frame for plotting.}

\item{y}{Character, specifying the percentage of each group.}

\item{fill}{Character, specifying the groups.}

\item{rows}{Character, specifying the row in the pie matrix.}

\item{cols}{Character, specifying the column in the pie matrix.}

\item{percentage}{Boolean specifying whether convert the frequency to percentage before plotting.}

\item{size}{Numeric, specifying the pie size.}

\item{row.label}{A named character vector, specifying the row labels.}

\item{col.label}{A named character vector, specifying the col labels.}
}
\description{
A matrix of pie plots
}
