% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylineView.R
\docType{methods}
\name{polylineView}
\alias{polylineView}
\title{polyline plot}
\usage{
polylineView(gg, x = 0, xlab = NULL, ylab = "Count", colab = NULL,
  main = NULL, filename = NULL, width = 5, height = 4, ...)
}
\arguments{
\item{gg}{Data frame}

\item{x}{Can be integer (variable position) or string (variable name).}

\item{xlab}{Label of x-axis}

\item{ylab}{Label of y-axis}

\item{colab}{Label of color}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Polyline plot for each column
}
\author{
Wubing Zhang
}
