% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-print.R
\name{print.plots_dims}
\alias{print.plots_dims}
\title{Print default canvas sizes}
\usage{
\method{print}{plots_dims}(x, ...)
}
\arguments{
\item{x}{Sizes, from \code{\link[=plots_dims]{plots_dims()}}.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Lists the canvas size each plot type gets by default.
}
\examples{
plots_dims(heatmap = c(18, 11))

}
