% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processArray.R
\docType{methods}
\name{processArray}
\alias{processArray}
\title{Method to preprocess Microarray dataset from GEO}
\usage{
processArray(expr = "GSE5821_series_matrix.txt", GPL = "GPL96.soft",
  symbol = NA, org = "hsa")
}
\arguments{
\item{expr}{A file path to GSE_series_matrix.}

\item{GPL}{A file path to GPL soft file.}

\item{symbol}{The colname of gene symbol in `GPL`.}

\item{org}{'hsa' or 'mmu'.}
}
\value{
A normalized expression profile.
}
\description{
Method to preprocess Microarray dataset from GEO
}
\author{
Wubing Zhang
}
