% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimma.R
\docType{methods}
\name{runLimma}
\alias{runLimma}
\title{Call differential expressed genes using limma}
\usage{
runLimma(mat, idx_ctrl, idx_treat, pvalueCutoff = 0.25)
}
\arguments{
\item{mat}{Matrix like object.}

\item{idx_ctrl}{Index of control samples.}

\item{idx_treat}{Index of Treatment samples.}

\item{pvalueCutoff}{Pvalue cutoff.}
}
\value{
A list contains limma result table and ggplot object (volcano figure).
}
\description{
Call differential expressed genes using limma
}
\author{
Wubing Zhang
}
