% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\docType{methods}
\name{selector}
\alias{selector}
\title{Select signatures from candidate list (according to the consistence in most samples).}
\usage{
selector(mat, cutoff = 0, type = "<", select = 0.8)
}
\arguments{
\item{mat}{Data matrix, each row is candidates (genes), each column is samples.}

\item{cutoff}{Cutoff to define the signatures.}

\item{type}{Direction to select signatures.}

\item{select}{Proportion of samples in which signature is selected.}
}
\value{
An list containing two elements, first is selected signature and second is a ggplot object.
}
\description{
Select signatures from candidate list (according to the consistence in most samples).
}
\author{
Wubing Zhang
}
