% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareDiff.R
\docType{methods}
\name{squareDiff}
\alias{squareDiff}
\title{Differential expression analysis of square groups.}
\usage{
squareDiff(expr, x = "TRAF3", y = c("TAP1", "B2M", "HLA-A", "HLA-B",
  "HLA-C"), x_cutoff = c(-0.5, 0.5), y_cutoff = c(-0.5, 0.5), ...)
}
\arguments{
\item{expr}{Matrix like object, each row is gene and each column is sample.}

\item{x}{Specifying x-axis which should be valid row names.}

\item{y}{Specifying y-axis which should be valid row names.}

\item{x_cutoff}{Specifying cutoff to classify x into high and low groups.}

\item{y_cutoff}{Specifying cutoff to classify x into high and low groups.}
}
\value{
A list.
}
\description{
Differential expression analysis of square groups.
}
\author{
Wubing Zhang
}
