% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareView.R
\docType{methods}
\name{SquareView}
\alias{SquareView}
\title{Scatter View of MHC-I and PD-L1 level, and separate all samples by 9 square.}
\usage{
SquareView(mat, x = "Ctrl.Mean", y = "IFNr.Mean", option = 1, label = 0,
  top = 10, genelist = c(), x_cutoff = c(-1, 1), y_cutoff = c(-1, 1),
  groups = c("DN", "MP_PN", "DP", "MN_PP"))
}
\arguments{
\item{mat}{Data frame, each row is gene and each column is sample.}

\item{x}{Specifying x-axis which should be valid col names.}

\item{y}{Specifying y-axis which should be valid col names.}

\item{option}{1 or 2.}

\item{label}{Specifying label text which should be valid col name.}

\item{top}{The number of top terms to be labeled.}

\item{genelist}{Interested term sets to be labeled.}

\item{x_cutoff}{Specifying cutoff to classify x into high and low groups.}

\item{y_cutoff}{Specifying cutoff to classify x into high and low groups.}

\item{groups}{Specifying name of each group.}
}
\value{
ggplot object.
}
\description{
Scatter View of MHC-I and PD-L1 level, and separate all samples by 9 square.
}
\author{
Wubing Zhang
}
