% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\usage{
summary(object)
}
\arguments{
\item{object}{A \code{ExprDataSet} instance.}

\item{...}{additional parameter}
}
\value{
A data frame
}
\description{
summary method for \code{ExprDataSet} instance
}
\author{
Wubing Zhang
}
