% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_interaction.R
\name{Survival_interaction}
\alias{Survival_interaction}
\title{Survival interaction in TIDE}
\usage{
Survival_interaction(mat, pivot, survival, CTL_genes = c("CD8A", "CD8B",
  "GZMA", "GZMB", "PRF1"))
}
\description{
Survival interaction in TIDE
}
