% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSeqs.R
\docType{methods}
\name{tableSeqs}
\alias{tableSeqs}
\title{Table creation for sequences}
\usage{
tableSeqs(sequences, len = nchar(sequences[1]))
}
\arguments{
\item{sequences}{A character vector with the same length of sequences.}

\item{len}{Integer, specifying the read length.}
}
\value{
A data frame.
}
\description{
Build a contingency table of the counts at each position of sequences.
}
\author{
Wubing Zhang
}
