% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIDE_Coculture_View.R
\docType{methods}
\name{TIDE_Coculture_View}
\alias{TIDE_Coculture_View}
\title{Visualize the selection level of gene in TIDE coculture screens.}
\usage{
TIDE_Coculture_View(gene = "Traf3",
  PanDir = "~/Jobs/Archive/CRISPRscreens/TIDE_coculture/Pan_2018_normcount.txt",
  MangusoDir1 = "~/Jobs/Archive/CRISPRscreens/TIDE_coculture/Manguso_2017/Manguso_lfc.txt",
  IgGDir = "~/Jobs/Archive/CRISPRscreens/TIDE_coculture/Kearney_2018/test/MC38_IgGvsT30.sgrna_summary.txt",
  PD1Dir = "~/Jobs/Archive/CRISPRscreens/TIDE_coculture/Kearney_2018/test/MC38_PD1vsT30.sgrna_summary.txt",
  NK10Dir = "~/Jobs/Archive/CRISPRscreens/TIDE_coculture/Kearney_2018/test/NK_Hit10vsTend.sgrna_summary.txt",
  NK20Dir = "~/Jobs/Archive/CRISPRscreens/TIDE_coculture/Kearney_2018/test/NK_Hit20vsTend.sgrna_summary.txt")
}
\arguments{
\item{gene}{A vector of gene names.}

\item{PanDir}{The path to Pan's screen.}

\item{MangusoDir1}{The path to Manguso's screen.}
}
\description{
Visualize the selection level of gene in TIDE coculture screens.
}
\examples{
cocultureView("Traf3", width = 6, height = 3)
}
