% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timer_DiffExpr_View.R
\docType{methods}
\name{Timer_DiffExpr_View}
\alias{Timer_DiffExpr_View}
\title{Visualize the differential expression results in TCGA data (Timer)}
\usage{
Timer_DiffExpr_View(gene, exprdir = "ProcessedDat/")
}
\arguments{
\item{gene}{A vector of gene names.}

\item{exprdir}{The directory of processed TCGA expression profiles.}
}
\description{
Visualize the differential expression results in TCGA data (Timer)
}
\examples{
args <- commandArgs(T)
if(length(args)==0 | args[1]=="-h"){
  message("Usage: Rscript Timer_DiffExpr_View.R gene outdir")
  q("no")
}
gene = args[1]
outdir = args[2]
}
