% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoView.R
\docType{methods}
\name{VolcanoView}
\alias{VolcanoView}
\title{Volcano View}
\usage{
VolcanoView(mat, x = "logFC", y = "nLogP", Label = NA, top = 0,
  topnames = NULL, filename = NULL, x_cutoff = 0.1,
  y_cutoff = -log10(0.05), main = NULL, xlab = "Log2 Fold Change",
  ylab = "-Log10(Adjust.P)", ...)
}
\arguments{
\item{mat}{Data frame}

\item{x}{Colname of mat specifying x-axis in Volcanno figure.}

\item{y}{Colname of mat specifying y-axis in Volcanno figure.}

\item{Label}{Colname of mat specifying labeled terms in Volcanno figure.}

\item{top}{Interger, label top significant terms.}

\item{topnames}{Character vector, indicating interested terms to be labeled.}

\item{filename}{Figure file name to create on disk. Default filename="NULL",
which means don't save the figure on disk.}

\item{x_cutoff}{Cutoff of x-axis.}

\item{y_cutoff}{Cutoff of y-axis.}

\item{main}{Title of volcano figure.}

\item{xlab}{Label of x-axis in figure.}

\item{ylab}{Label of t-axis in figure.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Volcano View
}
\author{
Wubing Zhang
}
