% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxView.R
\docType{methods}
\name{BoxView}
\alias{BoxView}
\title{Boxplot}
\usage{
BoxView(
  gg,
  x,
  y,
  color = NA,
  fill = NA,
  width = 0.6,
  size = 1,
  comparisons = NULL,
  test.method = "t.test",
  add.jitter = FALSE,
  jitter.color = color,
  jitter.size = size,
  alpha = 0.6,
  main = NULL,
  xlab = x,
  ylab = y,
  legend.position = "none",
  ...
)
}
\arguments{
\item{gg}{A data frame.}

\item{x}{Character, specifying the column name for x plotting.}

\item{y}{Character, specifying the column name for y plotting.}

\item{color}{Character, specifying the column name/color.}

\item{fill}{Character, specifying the column name/fill color.}

\item{width}{Numeric, specifying the box width.}

\item{size}{Numeric, specifying size of the box.}

\item{comparisons}{A list, customizing the comparisons.}

\item{test.method}{Character, specifying the statistical test method.}

\item{add.jitter}{Boolean, whether add jitter into the plot.}

\item{jitter.color}{Character, specifying the column name/color of the points.}

\item{jitter.size}{Numeric, specifying size of the jitter.}

\item{alpha}{A numeric, specifying the transparency of the dots.}

\item{main}{Title of the figure.}

\item{xlab}{Title of x-axis}

\item{ylab}{Title of y-axis.}

\item{legend.position}{Position of legend, "none", "right", "top", "bottom", or
a two-length vector indicating the position.}

\item{...}{Other parameters in geom_boxplot.}
}
\value{
An object created by \code{ggplot}, which can be assigned
and further customized.
}
\description{
Boxplot
}
\author{
Wubing Zhang
}
