% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOrgGeneList.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Basic function to convert mouse to human gene names}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{A vector of gene symbols.}
}
\description{
Basic function to convert mouse to human gene names
}
