% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalyze.R
\docType{methods}
\name{DEAnalyze}
\alias{DEAnalyze}
\title{Differential expression analysis}
\usage{
DEAnalyze(obj, SampleAnn = NULL, type = "Array", minS = 10,
  trans.method = "vst")
}
\arguments{
\item{obj}{Matrix like object or an ExprDataSet instance.}

\item{SampleAnn}{Matrix like object (only when obj is a matrix),
the rownames should match colnames in obj, and the first column should be Condition.}

\item{type}{"Array" or "RNASeq", only needed when obj is matrix like object.}

\item{minS}{Same as that in `TransformCount`.}

\item{trans.method}{Same as `method` in `TransformCount`.}
}
\value{
An ExprDataSet instance.
}
\description{
Differential expression analysis
}
\seealso{
\code{\link{ExprDataSet-class}}
}
\author{
Wubing Zhang
}
