% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapView.R
\docType{methods}
\name{heatmapView}
\alias{heatmapView}
\title{Draw heatmap}
\usage{
heatmapView(mat, limit = 2, cluster_rows = FALSE, cluster_cols = FALSE,
  filename = NA, width = NA, height = NA, ...)
}
\arguments{
\item{mat}{Matrix like object, each row is gene and each column is sample.}

\item{limit}{Max value in heatmap}

\item{filename}{File path where to save the picture.}

\item{width}{Manual option for determining the output file width in inches.}

\item{height}{Manual option for determining the output file height in inches.}
}
\value{
Invisibly a pheatmap object that is a list with components.
}
\description{
Draw heatmap
}
\author{
Wubing Zhang
}
