% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMView.R
\name{KMView}
\alias{KMView}
\title{Survival analysis and plot the kaplan-meier curves}
\usage{
KMView(survdata, bio, os = "os", event = "event", cut = c(0, 0.5, 1),
  optimalCut = TRUE, pval.method = c("cox", "logrank"),
  pval.pos = c(0.9, 0.95), labels = 1:(length(cut) - 1), ...)
}
\arguments{
\item{survdata}{A data frame including the clinical information.}

\item{bio}{A character or an integer, specifying which variable/column
in the survdata should be used to predict the risk of patients.}

\item{os}{A character or an integer, specifying the variable/column of overall survival.}

\item{event}{A character or an integer, specifying the variable/column of outcome.}

\item{cut}{A numeric vector for separating the patients into different groups,
only available when `bio` is a continuous variable.}

\item{optimalCut}{Boolean, indicating whether determin the ootimal cutpoint.}

\item{pval.method}{Character, indicating the method for the pvalue calculation,
only logrank and cox are available now.}

\item{pval.pos}{Character, indicating the method for the pvalue calculation.}

\item{labels}{A character vector, specifying the labels for each patient group.}

\item{...}{Other available parameters in ggsurvplot.}

\item{event}{A character or an integer, specifying the variable/column of outcome.}
}
\value{
A ggplot instance that allows further custimization.
}
\description{
Survival analysis and plot the kaplan-meier curves
}
\author{
Wubing Zhang
}
